Create Convert.cpp
Had trouble working with the original main.cpp, so I changed it over to something simpler. I don't believe tuples are needed as our input is always formatted with the first four lines describing the nfa, then the rest listing the transitions. We can just look at each line one at a time to fill the table. Almost done sorting the input for the NFA table, then I'll transform it to DFA and it's done.